+++ Data/Invoice.cs
@@ -9,26 +9,26 @@ namespace Exercise_03.Data
 {
     public class Invoice
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Display(Name = "Party")]
         [Required]
         //[ForeignKey("Party")]
-        public int partyId { get; set; }
+        public long partyId { get; set; }
 
         [Display(Name = "Product")]
         [Required]
         //[ForeignKey("Product")]
-        public int productId { get; set; }
+        public long productId { get; set; }
 
         [Required]
         [Display(Name = "Current Rate")]
-        public int CurrentRate { get; set; }
+        public long CurrentRate { get; set; }
 
         [Required]
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
 
-        public int Total { get; set; }
+        public long Total { get; set; }
 
 
         //public Party Party { get; set; }
