+++ Models/InvoiceModel.cs
@@ -9,12 +9,12 @@ namespace Exercise_03.Models
 {
     public class InvoiceModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Display(Name = "Party")]
         [Required]
         //[ForeignKey("Party")]
-        public int partyId { get; set; }
+        public long partyId { get; set; }
 
      
         public string partyName { get; set; }
@@ -25,18 +25,18 @@ namespace Exercise_03.Models
         [Display(Name = "Product")]
         [Required]
         //[ForeignKey("Product")]
-        public int productId { get; set; }
+        public long productId { get; set; }
 
         [Required]
         [Display(Name = "Current Rate")]
        
-        public int CurrentRate { get; set; }
+        public long CurrentRate { get; set; }
 
         [Required]
         [Range(1,100)]
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
 
-        public int Total { get; set; }
+        public long Total { get; set; }
 
         //public Party Party { get; set; }
